Conflicts:
	target/m2e-wtp/web-resources/META-INF/maven/cl.awakelab/modulo5dia12/pom.properties

+++ src/main/java/cl/m5d12/servicio/CapacitacionImple.java
@@ -16,9 +16,9 @@ public class CapacitacionImple implements CapacitacionServicio {
 	CapacitacionRepository caprepo;
 	
 	@Override
-	public Capacitacion findCapacitacionByid(int idCapa) {
+	public Capacitacion findCapacitacionByid(long idCapa) {
 		
-		return caprepo.findOne(idCapa);
+		return caprepo.findOne((int) idCapa);
 	}
 
 	@Override
